pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tav"
version = "0.1.0"
description = "Topology-Aware Vector Routing for documents"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    "pymupdf>=1.26.4",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
app = ["streamlit>=1.30.0"]
postgres = ["psycopg[binary]>=3.1.0", "pgvector>=0.2.0"]
mongo = ["pymongo>=4.6.0"]

[project.scripts]
tav = "tav.cli:main"

[tool.setuptools.packages.find]
include = ["tav*"]
exclude = ["tests*"]
